+++ Domain/Entidades/Estudiante.cs
@@ -15,8 +15,8 @@ namespace Domain.Entidades
         public int LugarEntreHermanos { get; private set; }
         public string SeguroSocial { get; private set; }
         public float PuntajeSisben { get; private set; }
-        public int MatriculaId { get; private set; }
-        public int? CursoId { get; private set; }
+        public long MatriculaId { get; private set; }
+        public long? CursoId { get; private set; }
         public Acudiente RepresentanteLegal { get; private set; }
         public PensionEscolar PensionEscolar { get; private set; }
         public List<Nota> ListaNotas { get; set; }
